+++ .eslintrc.js
@@ -47,7 +47,8 @@ module.exports = {
     ],
     'react/prop-types': 0,
     '@typescript-eslint/explicit-function-return-type': 0,
-    '@typescript-eslint/no-explicit-any': 0
+    '@typescript-eslint/no-explicit-any': 0,
+    '@typescript-eslint/member-delimiter-style': 0
   },
   settings: {
     react: {