+++ packages/rum-core/src/common/constants.js
@@ -100,6 +100,7 @@ const CONFIG_CHANGE = 'config:change'
 const XMLHTTPREQUEST = 'xmlhttprequest'
 const FETCH = 'fetch'
 const HISTORY = 'history'
+const EVENT_TARGET = 'eventtarget'
 const ERROR = 'error'
 
 /**
@@ -112,6 +113,18 @@ const AFTER_EVENT = ':after'
  * Local Config Key used storing the remote config in the localStorage
  */
 const LOCAL_CONFIG_KEY = 'elastic_apm_config'
+
+/**
+ * List of entry types that could be observed by the PerformaneObserver Interface
+ * or can be captured via Performance Timeline API
+ */
+const LONG_TASK = 'longtask'
+const PAINT = 'paint'
+const MEASURE = 'measure'
+const NAVIGATION = 'navigation'
+const RESOURCE = 'resource'
+const FIRST_CONTENTFUL_PAINT = 'first-contentful-paint'
+const LARGEST_CONTENTFUL_PAINT = 'largest-contentful-paint'
 
 /**
  * Default configs used on top of extensible configs from ConfigService
@@ -142,11 +155,19 @@ export {
   XMLHTTPREQUEST,
   FETCH,
   HISTORY,
+  EVENT_TARGET,
   ERROR,
   BEFORE_EVENT,
   AFTER_EVENT,
   LOCAL_CONFIG_KEY,
   HTTP_REQUEST_TYPE,
+  LONG_TASK,
+  PAINT,
+  MEASURE,
+  NAVIGATION,
+  RESOURCE,
+  FIRST_CONTENTFUL_PAINT,
+  LARGEST_CONTENTFUL_PAINT,
   KEYWORD_LIMIT,
   SERVER_URL_PREFIX,
   BROWSER_RESPONSIVENESS_INTERVAL,