+++ packages/rum-core/src/common/patching/index.js
@@ -26,8 +26,10 @@
 import { patchXMLHttpRequest } from './xhr-patch'
 import { patchFetch } from './fetch-patch'
 import { patchHistory } from './history-patch'
+import { patchEventTarget } from './event-target-patch'
 import EventHandler from '../event-handler'
-import { HISTORY, FETCH, XMLHTTPREQUEST } from '../constants'
+
+import { HISTORY, FETCH, XMLHTTPREQUEST, EVENT_TARGET } from '../constants'
 
 const patchEventHandler = new EventHandler()
 
@@ -43,6 +45,10 @@ function patchAll() {
     })
     patchHistory(function(event, task) {
       patchEventHandler.send(HISTORY, [event, task])
+    })
+
+    patchEventTarget(function(event, task) {
+      patchEventHandler.send(EVENT_TARGET, [event, task])
     })
   }
   return patchEventHandler