+++ packages/rum-core/src/error-logging/error-logging.js
@@ -32,6 +32,43 @@ import {
 } from '../common/utils'
 import { truncateModel, ERROR_MODEL } from '../common/truncate'
 
+/**
+ * List of keys to be ignored from getting added to custom error properties
+ */
+const IGNORE_KEYS = ['stack', 'message']
+
+function getErrorProperties(error) {
+  /**
+   * Flag which is used to eliminate the empty object
+   * check on context.custom
+   */
+  let propertyFound = false
+  const properties = {}
+  Object.keys(error).forEach(function(key) {
+    if (IGNORE_KEYS.indexOf(key) >= 0) {
+      return
+    }
+    /**
+     * ignore null, undefined, function values
+     */
+    let val = error[key]
+    if (val == null || typeof val === 'function') {
+      return
+    }
+
+    if (typeof val === 'object') {
+      if (typeof val.toISOString !== 'function') return
+      val = val.toISOString()
+    }
+    properties[key] = val
+    propertyFound = true
+  })
+
+  if (propertyFound) {
+    return properties
+  }
+}
+
 class ErrorLogging {
   constructor(apmServer, configService, transactionService) {
     this._apmServer = apmServer
@@ -60,7 +97,10 @@ class ErrorLogging {
     if (error && typeof error === 'object') {
       errorMessage = errorMessage || error.message
       errorType = error.name
-      errorContext = this._getErrorProperties(error)
+      const customProperties = getErrorProperties(error)
+      if (customProperties) {
+        errorContext.custom = customProperties
+      }
     }
 
     if (!errorType) {
@@ -163,25 +203,6 @@ class ErrorLogging {
       errorEvent.error = messageOrError
     }
     return this.logErrorEvent(errorEvent)
-  }
-
-  _getErrorProperties(error) {
-    const properties = {}
-    Object.keys(error).forEach(function(key) {
-      if (key === 'stack') return
-      let val = error[key]
-      if (val === null) return // null is typeof object and well break the switch below
-      switch (typeof val) {
-        case 'function':
-          return
-        case 'object':
-          // ignore all objects except Dates
-          if (typeof val.toISOString !== 'function') return
-          val = val.toISOString()
-      }
-      properties[key] = val
-    })
-    return properties
   }
 }