+++ packages/rum-vue/src/index.js
@@ -25,18 +25,32 @@
 
 import { apmBase } from '@elastic/apm-rum'
 import { routeHooks } from './route-hooks'
+import { getErrorHandler } from './error-handler'
 
 export const ApmVuePlugin = {
   install: (Vue, options) => {
-    const { router, apm = apmBase, config } = options
+    const { router, apm = apmBase, config, captureErrors = true } = options
     /**
      * Initialize the APM with the config
      */
     apm.init(config)
 
-    routeHooks(router, apm)
     /**
-     * Provide it via $apm to be accessed in all Vue Components
+     * Hook router if provided
+     */
+    if (router) {
+      routeHooks(router, apm)
+    }
+
+    if (captureErrors) {
+      /**
+       * Global error handler for capturing errors during
+       * component renders
+       */
+      Vue.config.errorHandler = getErrorHandler(Vue, apm)
+    }
+    /**
+     * Provide the APM instance via $apm to be accessed in all Vue Components
      */
     Vue.prototype.$apm = apm
   }