+++ packages/rum/src/apm-base.js
@@ -43,7 +43,7 @@ class ApmBase {
       /**
        * Set Agent version to be sent as part of metadata to the APM Server
        */
-      configService.setVersion('4.6.0')
+      configService.setVersion('4.7.1')
       this.config(config)
       const loggingService = this.serviceFactory.getService('LoggingService')
       /**
@@ -79,6 +79,7 @@ class ApmBase {
           sendPageLoad()
         }
       } else {
+        this._disable = true
         loggingService.info('RUM agent is inactive')
       }
     }